requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pyodbc==5.2.0
python-dotenv==1.1.1
azure-servicebus==7.12.3
